requirements.txt
torch>=2.0.0
transformers>=4.46.0
accelerate
bitsandbytes
datasets
wandb
Levenshtein
Pillow
tqdm
pandas
timm
torchvision
pyyaml
